Complete cleanup: Remove all JSON snapshot columns and fix type consistency
Database Migration (V6):
- Removed delivery_address_snapshot column (was TEXT/JSON)
- Removed delivery_option_snapshot column (was TEXT/JSON)
- Removed pricing_before_snapshot column (was TEXT/JSON)
- Removed pricing_after_snapshot column (was TEXT/JSON)
- Removed order_discount_metadata column (was TEXT/JSON)
- Removed before_snapshot column from order_items (was TEXT/JSON)
- Removed after_snapshot column from order_items (was TEXT/JSON)
- Removed audit_metadata column from order_items (was TEXT/JSON)
- Kept all new proper typed columns

Entity Models:
- Order.java: Removed deliveryAddressSnapshot, deliveryOptionSnapshot, pricingBeforeSnapshot, pricingAfterSnapshot, orderDiscountMetadata
- Order.java: Fixed deliveryLatitude, deliveryLongitude to BigDecimal (matches request DTOs)
- OrderItem.java: Removed beforeSnapshot, afterSnapshot, auditMetadata

Helper DTOs:
- OrderCreateHelper.java: Fixed deliveryLatitude, deliveryLongitude to BigDecimal

Result:
✅ ZERO JSON storage in entire order feature
✅ All data in typed database columns (BigDecimal, LocalDateTime, String)
✅ Type-consistent throughout: BigDecimal for coordinates, prices, amounts
✅ Clean database schema with no deprecated columns
✅ Ready for production use

+++ menu-scanner-backend/src/main/java/com/emenu/features/order/dto/helper/OrderCreateHelper.java
@@ -35,8 +35,8 @@ public class OrderCreateHelper {
     private String deliveryStreetNumber;
     private String deliveryHouseNumber;
     private String deliveryNote;
-    private Double deliveryLatitude;
-    private Double deliveryLongitude;
+    private BigDecimal deliveryLatitude;
+    private BigDecimal deliveryLongitude;
 
     // ===== Delivery Option Fields =====
     private String deliveryOptionName;
